Make quantity field in PR Request and amount in Voucher allow decimal values

+++ EDocument_Data/DTOs/Requests/DiscountRequest/ApproveMultimediaRequestDto.cs
@@ -10,7 +10,7 @@ namespace EDocument_Data.DTOs.Requests.DiscountRequest
     {
         [Required]
         public long RequestId { get; init; }
-        public float? DiscountAmount { get; init; }
+        public double? DiscountAmount { get; init; }
         [FileValidation(100 * 1024 * 1024, true, ".pdf", ".docx", ".xlsx", ".xls", ".csv", ".zip", ".png", ".jpg", ".jpeg")]
         public IFormFile? HoSupportedDocument { get; init; }
         public string? ReviewedNotes { get; init; }
