+++ EDocument_Data/DTOs/Requests/DiscountRequest/DiscountRequestReadDto.cs
@@ -35,7 +35,7 @@ namespace EDocument_Data.DTOs.Requests.DiscountRequest
         public string RelatedContainers { get; set; }
         public string DiscountType { get; set; }
         public string DiscountReason { get; set; }
-        public float DiscountAmount { get; set; }
+        public double DiscountAmount { get; set; }
 
         public AttachmentReadDto HoSupportedDocument { get; set; }
         public List<AttachmentReadDto> Attachments { get; set; }
