+++ EDocument_Data/Models/DiscountRequest.cs
@@ -20,7 +20,7 @@ namespace EDocument_Data.Models
         public string RelatedContainers { get; set; }
         public string DiscountType { get; set; }
         public string DiscountReason { get; set; }
-        public float DiscountAmount { get; set; }
+        public double DiscountAmount { get; set; }
         public string HoSupportedDocumentPath { get; set; }
         public Request Request { get; set; }
     }
